Modify getUsersCount method in UserDao

+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/persistence/UserDao.java
@@ -19,7 +19,7 @@ public interface UserDao {
 
     void changePassword(String email, String password);
 
-    Integer getUsersCount();
+    long getUsersCount();
 
     Integer getUsersCountByFilters(String categoryId, String location, String educationLevel);
