+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/UserService.java
@@ -20,7 +20,7 @@ public interface UserService {
 
     void changePassword(String email, String password);
 
-    Integer getUsersCount();
+    long getUsersCount();
 
     Integer getUsersCountByFilters(String categoryId, String location, String educationLevel);
