+++ paw-webapp/services/src/main/java/ar/edu/itba/paw/services/UserServiceImpl.java
@@ -63,7 +63,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Integer getUsersCount() {
+    public long getUsersCount() {
         return userDao.getUsersCount();
     }
